PrivacyViolation.java
Privacy Violation                                                         

Risk                                                                      

What might happen                                                         

A users personal information could be stolen by a malicious programmer,  

or an attacker that intercepts the data.                                  

                                                                          

Cause                                                                     

How does it happen                                                        

The application sends user information, such as passwords, account        

information, or credit card numbers, outside the application, such        

as writing it to a local text or log file or sending it to an external    

web service.                                                              

                                                                          

General Recommendations                                                   

How to avoid it                                                           

1. Personal data should be removed before writing to logs or other files. 

2. Review the need and justification of sending personal data to remote   

web services.