StoredXSS.java
Stored XSS                                                                                 

                                                                                           

Risk                                                                                       

What might happen                                                                          

An attacker could use legitimate access to the application to submit engineered data       

to the applications database. When another user subsequently accesses this data,          

web pages may be rewritten and malicious scripts may be activated.                         

                                                                                           

Cause                                                                                      

How does it happen                                                                         

The application creates web pages that include data from the applications database.       

The data is embedded directly in the page's HTML, causing the browser to display it        

as part of the web page. This data may have originated in input from another user.         

If the data includes HTML fragments or Javascript, these are displayed too, and the user   

cannot tell that this is not the intended page. The vulnerability is the result of         

embedding arbitrary database data without first encoding it in a format that would         

prevent the browser from treating it like HTML instead of plain text.                      

                                                                                           

General Recommendations                                                                    

How to avoid it                                                                            

1. Validate all dynamic data, regardless of source. Validation should be based on          

a whitelist: accept only data fitting a specified structure, rather than reject            

bad patterns. Check for:                                                                   

		o Data type                                                                           

		o Size                                                                                

		o Range                                                                               

		o Format                                                                              

		o Expected values                                                                     

2. Validation is not a replacement for encoding. Fully encode all dynamic data,            

regardless of source, before embedding it in output. Encoding should be context-sensitive. 

For example:                                                                               

		o HTML encoding for HTML content                                                      

		o HTML attribute encoding for data output to attribute values                         

		o Javascript encoding for server-generated Javascript.                                

3. Consider using either the ESAPI encoding library, or its built-in functions. For        

earlier versions of ASP.NET, consider using the AntiXSS library.                           

4. In the Content-Type HTTP response header, explicitly define character encoding (charset)

for the entire page.                                                                       

5. Set the httpOnly flag on the session cookie, to prevent XSS exploits from stealing      

the cookie.